# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Qapac API - Transporte Publico de Cajamarca
  description: |
    API REST para la aplicacion movil de transporte publico en Cajamarca, Peru.

    ## Audiencia
    Equipo de desarrollo movil (Android) que consume esta API.

    ## Autenticacion
    - Los endpoints **publicos** no requieren autenticacion (usuarios anonimos).
    - Los endpoints de **conductor** y **admin** requieren un JWT en el header `Authorization: Bearer <token>`.
    - Los tokens se obtienen via `POST /api/v1/auth/login`.
    - El access token tiene un TTL de 15 minutos; el refresh token 7 dias.
    - Se usa rotacion de refresh tokens: cada refresh emite un nuevo par.

    ## Formato de errores
    Todos los errores retornan:
    ```json
    {"error": "<mensaje>"}
    ```

    ## Polling de posiciones en tiempo real
    No hay WebSocket. El cliente movil debe hacer polling periodico a los endpoints
    de posicion de vehiculos (ej. cada 5-10 segundos).
  version: 1.0.0
  contact:
    name: Equipo Backend Qapac
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.qapac.pe
    description: Servidor de produccion (futuro)
...
paths:
...
  /api/v1/routes:
    get:
      tags: [Rutas]
      summary: Listar todas las rutas
      description: |
        Retorna todas las rutas de transporte con la cantidad de vehiculos
        activos en cada una. No requiere autenticacion.
      operationId: listRoutes
      responses:
        "200":
          description: Lista de rutas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteBasic"
              example:
                - id: 1
                  name: "R01 - Centro a Baños del Inca"
                  active: true
                  vehicle_count: 3
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes/{id}:
    get:
      tags: [Rutas]
      summary: Obtener detalle de una ruta
      description: |
        Retorna la ruta con sus paradas ordenadas, vehiculos asignados,
        y la polyline del recorrido.
      operationId: getRoute
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Detalle de la ruta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Ruta no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "route not found"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes/{id}/vehicles:
    get:
      tags: [Rutas]
      summary: Vehiculos de una ruta con posicion GPS
      description: |
        Retorna todos los vehiculos asignados a una ruta, incluyendo su
        posicion GPS mas reciente (si existe). Ideal para mostrar buses
        en el mapa en tiempo real.
      operationId: getRouteVehicles
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Lista de vehiculos con posicion
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteVehicleWithPosition"
              example:
                - id: 1
                  plate: "ABC-123"
                  driver_name: "Carlos Quispe"
                  collector_name: "Maria Lopez"
                  status: active
                  position:
                    lat: -7.1640
                    lon: -78.5010
                    heading: 180.5
                    speed: 25.3
                    recorded_at: "2025-01-15T10:30:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/vehicles/nearby:
    get:
      tags: [Vehiculos]
      summary: Buscar vehiculos cercanos
      description: |
        Retorna los vehiculos que tienen posicion GPS registrada dentro del
        radio especificado. No requiere autenticacion.
      operationId: nearbyVehicles
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitud WGS-84
          schema:
            type: number
            format: double
            example: -7.1637
        - name: lon
          in: query
          required: true
          description: Longitud WGS-84
          schema:
            type: number
            format: double
            example: -78.5003
        - name: radius
          in: query
          required: false
          description: "Radio en metros (default: 1000, max: 50000)"
          schema:
            type: number
            format: double
            default: 1000
            maximum: 50000
      responses:
        "200":
          description: Lista de vehiculos cercanos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyVehicle"
              example:
                - id: 1
                  plate: "ABC-123"
                  route_name: "R01 - Centro a Baños del Inca"
                  lat: -7.1640
                  lon: -78.5010
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/vehicles/{id}/position:
    get:
      tags: [Vehiculos]
      summary: Obtener posicion GPS de un vehiculo
      description: |
        Retorna la ultima posicion GPS registrada de un vehiculo especifico.
      operationId: getVehiclePosition
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Posicion del vehiculo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehiclePosition"
              example:
                lat: -7.1640
                lon: -78.5010
                heading: 180.5
                speed: 25.3
                recorded_at: "2025-01-15T10:30:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: No hay posicion registrada para este vehiculo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "no position recorded for this vehicle"
        "500":
          $ref: "#/components/responses/InternalError"

  # ===========================================================================
  # Alertas - Publico (lectura)
  # ===========================================================================
  /api/v1/alerts:
    get:
      tags: [Alertas]
      summary: Listar alertas de servicio
      description: |
        Retorna las alertas activas. Opcionalmente se puede filtrar por ruta.
        No requiere autenticacion.
      operationId: listAlerts
      parameters:
        - name: route_id
          in: query
          required: false
          description: Filtrar alertas por ID de ruta
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/alerts/{id}:
    get:
      tags: [Alertas]
      summary: Obtener detalle de una alerta
      operationId: getAlert
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Detalle de la alerta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Alerta no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "alert not found"
        "500":
          $ref: "#/components/responses/InternalError"

  # ===========================================================================
  # Calificaciones - Publico
  # ===========================================================================
  /api/v1/ratings:
    post:
      tags: [Calificaciones]
      summary: Calificar un viaje
      description: |
        Permite a usuarios anonimos calificar un viaje (1-5 estrellas).
        Se identifica al usuario por `device_id` (UUID del dispositivo Android).
        Un dispositivo solo puede calificar un viaje una vez (409 si se repite).
      operationId: createRating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingRequest"
            example:
              trip_id: 1
              rating: 4
              device_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "201":
          description: Calificacion creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
              example:
                id: 1
                trip_id: 1
                rating: 4
                device_id: "550e8400-e29b-41d4-a716-446655440000"
                created_at: "2025-01-15T10:30:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: El dispositivo ya califico este viaje
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "device has already rated this trip"
        "500":
          $ref: "#/components/responses/InternalError"

  # ===========================================================================
  # Favoritos - Publico
  # ===========================================================================
  /api/v1/favorites:
    get:
      tags: [Favoritos]
      summary: Listar rutas favoritas
      description: |
        Retorna las rutas favoritas de un dispositivo. Se identifica al
        usuario por `device_id` en query parameter.
      operationId: listFavorites
      parameters:
        - name: device_id
          in: query
          required: true
          description: UUID del dispositivo Android
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Lista de favoritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
              example:
                - id: 1
                  device_id: "550e8400-e29b-41d4-a716-446655440000"
                  route_id: 1
                  route_name: "R01 - Centro a Baños del Inca"
                  created_at: "2025-01-15T08:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Favoritos]
      summary: Agregar ruta a favoritos
      operationId: addFavorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteRequest"
            example:
              device_id: "550e8400-e29b-41d4-a716-446655440000"
              route_id: 1
      responses:
        "201":
          description: Favorito agregado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteCreated"
              example:
                id: 1
                device_id: "550e8400-e29b-41d4-a716-446655440000"
                route_id: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Favoritos]
      summary: Eliminar ruta de favoritos
      operationId: removeFavorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveFavoriteRequest"
            example:
              device_id: "550e8400-e29b-41d4-a716-446655440000"
              route_id: 1
      responses:
        "204":
          description: Favorito eliminado (sin cuerpo)
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ===========================================================================
  # Subida de archivos
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido via POST /api/v1/auth/login.
        Header: `Authorization: Bearer <access_token>`
  parameters:
    PathID:
      name: id
      in: path
      required: true
      description: Identificador numerico del recurso
      schema:
        type: integer
        format: int32
        minimum: 1
  responses:
    BadRequest:
      description: Solicitud mal formada o parametros invalidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje descriptivo del error

    # --- Auth ---
...
    RouteBasic:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        active:
          type: boolean
        vehicle_count:
          type: integer
          description: Cantidad de vehiculos activos en la ruta
    RouteDetail:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        active:
          type: boolean
        stops:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int32
              name:
                type: string
              lat:
                type: number
                format: double
              lon:
                type: number
                format: double
              sequence:
                type: integer
                description: Orden de la parada en la ruta
        vehicles:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int32
              plate:
                type: string
              driver:
                type: string
                nullable: true
              collector:
                type: string
                nullable: true
              status:
                type: string
        shape_polyline:
          type: string
          nullable: true
          description: "Polyline codificada del recorrido de la ruta (Google Encoded Polyline)"
    RouteVehicleWithPosition:
      type: object
      properties:
        id:
          type: integer
          format: int32
        plate:
          type: string
        driver_name:
          type: string
          nullable: true
        collector_name:
          type: string
          nullable: true
        status:
          type: string
        position:
          type: object
          nullable: true
          description: "Nulo si el vehiculo no tiene posicion GPS registrada"
          properties:
            lat:
              type: number
              format: double
            lon:
              type: number
              format: double
            heading:
              type: number
              format: double
              nullable: true
              description: "Direccion en grados (0-360)"
            speed:
              type: number
              format: double
              nullable: true
              description: "Velocidad en km/h"
            recorded_at:
              type: string
              format: date-time
...
    NearbyVehicle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        plate:
          type: string
        route_name:
          type: string
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
    VehiclePosition:
      type: object
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        heading:
          type: number
          format: double
          nullable: true
        speed:
          type: number
          format: double
          nullable: true
        recorded_at:
          type: string
          format: date-time

    # --- Alertas ---
    Alert:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        route_id:
          type: integer
          format: int32
          nullable: true
        vehicle_plate:
          type: string
        image_path:
          type: string
        created_by:
          type: integer
          format: int32
          nullable: true
        created_at:
          type: string
          format: date-time

    # --- Calificaciones ---
    CreateRatingRequest:
      type: object
      required: [trip_id, rating, device_id]
      properties:
        trip_id:
          type: integer
          format: int32
          description: ID del viaje a calificar
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: "Estrellas (1 a 5)"
        device_id:
          type: string
          description: UUID del dispositivo Android
    Rating:
      type: object
      properties:
        id:
          type: integer
          format: int32
        trip_id:
          type: integer
          format: int32
        rating:
          type: integer
        device_id:
          type: string
        created_at:
          type: string
          format: date-time

    # --- Favoritos ---
    Favorite:
      type: object
      properties:
        id:
          type: integer
          format: int32
        device_id:
          type: string
        route_id:
          type: integer
          format: int32
        route_name:
          type: string
        created_at:
          type: string
          format: date-time
    AddFavoriteRequest:
      type: object
      required: [device_id, route_id]
      properties:
        device_id:
          type: string
        route_id:
          type: integer
          format: int32
    RemoveFavoriteRequest:
      type: object
      required: [device_id, route_id]
      properties:
        device_id:
          type: string
        route_id:
          type: integer
          format: int32
    FavoriteCreated:
      type: object
      properties:
        id:
          type: integer
          format: int32
        device_id:
          type: string
        route_id:
          type: integer
          format: int32

    # --- Subida de archivos ---
...